Renamed Repo to CRUDInterface and added manager to maintain table size at max 5 (removes oldest entry when max size is reached then adds new one)
