+++ src/main/java/edoardosella/WeatherAPI/JPA/Repositories/PreviousRoutes/PreviousRoutesCRUDInterface.java
@@ -1,11 +1,12 @@
-package edoardosella.WeatherAPI.JPA.Repositories;
+package edoardosella.WeatherAPI.JPA.Repositories.PreviousRoutes;
 
 import edoardosella.WeatherAPI.JPA.Models.PreviousRoute;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface PreviousRouteRepo extends CrudRepository<PreviousRoute, Long> {
+public interface PreviousRoutesCRUDInterface extends CrudRepository<PreviousRoute, Long> {
     //https://stackoverflow.com/questions/27567351/get-last-records-ordered-by-data-on-spring-data
     List<PreviousRoute> findTop5ByOrderByIdDesc();
+    PreviousRoute findTopByOrderByIdAsc();
 }
